Extract FileMisc out of WorkspaceRegistry.
